+++ verify.c
@@ -192,8 +192,7 @@ int dd_verify(void)
     
             if (probe == PROBE_NONE || probed_size == 0)
                 fprintf(stderr, "\r%llu blocks (%lluMb) written.", 
-			/* [FIX] verify.c:195:25: warning: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument {3,4} has type ‘uintmax_t’ [-Wformat=] */
-                        (long long unsigned int) w_full, (long long unsigned int) total_mb);
+                        w_full, total_mb);
             else {
                 time_t curr_time = time(NULL);
                 int seconds = (int)difftime(curr_time, start_time);
@@ -208,8 +207,7 @@ int dd_verify(void)
                 time_left(secstr, sizeof secstr, seconds_remaining);
                 fprintf(stderr,
                         "\r[%d%% of %lluMb] %llu blocks (%lluMb) written. %s",
-			/* [FIX] verify.c:210:25: warning: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument {4,5,6} has type ‘off_t’ [-Wformat=] */
-                        prcnt, (long long unsigned int) probed_mb, (long long unsigned int) w_full, (long long unsigned int) total_mb, secstr);
+                        prcnt, probed_mb, w_full, total_mb, secstr);
             }	
         }
